fix: asset urls should not be stored with serverURL in the DB
